+++ src/attributecode/model.py
@@ -344,8 +344,7 @@ class ListField(StringField):
         if sval == oval:
             return True
 
-
-class UrlField(ListField):
+class UrlListField(ListField):
     """
     A URL field. The validated value is a list of URLs.
     """
@@ -353,13 +352,39 @@ class UrlField(ListField):
         """
         Check that URLs are valid. Return a list of errors.
         """
-        errors = super(UrlField, self)._validate(*args, ** kwargs)
-        for url in self.value:
+        errors = super(UrlListField, self)._validate(*args, ** kwargs)
+        name = self.name
+        val = self.value
+        for url in val:
             if not self.is_valid_url(url):
-                name = self.name
-                val = self.value
                 msg = (u'Field %(name)s: Invalid URL: %(val)s' % locals())
                 errors.append(Error(WARNING, msg))
+        return errors
+
+    @staticmethod
+    def is_valid_url(url):
+        """
+        Return True if a URL is valid.
+        """
+        scheme, netloc, _path, _p, _q, _frg = urlparse(url)
+        valid = scheme in ('http', 'https', 'ftp') and netloc
+        return valid
+
+
+class UrlField(StringField):
+    """
+    A URL field. The validated value is a URL.
+    """
+    def _validate(self, *args, **kwargs):
+        """
+        Check that URL is valid. Return a list of errors.
+        """
+        errors = super(UrlField, self)._validate(*args, ** kwargs)
+        name = self.name
+        val = self.value
+        if not self.is_valid_url(val):
+            msg = (u'Field %(name)s: Invalid URL: %(val)s' % locals())
+            errors.append(Error(WARNING, msg))
         return errors
 
     @staticmethod
@@ -692,7 +717,7 @@ class About(object):
             ('license_key', ListField()),
             ('license_name', ListField()),
             ('license_file', FileTextField()),
-            ('license_url', UrlField()),
+            ('license_url', UrlListField()),
             ('copyright', StringField()),
             ('notice_file', FileTextField()),
             ('notice_url', UrlField()),
@@ -704,7 +729,7 @@ class About(object):
 
             ('changelog_file', FileTextField()),
 
-            ('owner', ListField()),
+            ('owner', StringField()),
             ('owner_url', UrlField()),
             ('contact', ListField()),
             ('author', ListField()),
@@ -1066,7 +1091,7 @@ class About(object):
         return errors
 
 
-    def dumps(self, with_absent=False, with_empty=True):
+    def dumps(self, use_mapping=False, mapping_file=False, with_absent=False, with_empty=True):
         """
         Return self as a formatted ABOUT string.
         If with_absent, include absent (not present) fields.
@@ -1113,9 +1138,10 @@ class About(object):
             if lic_group[3]:
                 lic_dict['url'] = lic_group[3] 
             about_data.setdefault('licenses', []).append(lic_dict)
-        return saneyaml.dump(about_data)
+        formatted_about_data = util.format_output(about_data, use_mapping, mapping_file)
+        return saneyaml.dump(formatted_about_data)
 
-    def dump(self, location, with_absent=False, with_empty=True):
+    def dump(self, location, use_mapping=False, mapping_file=False, with_absent=False, with_empty=True):
         """
         Write formatted ABOUT representation of self to location.
         If with_absent, include absent (not present) fields.
@@ -1135,7 +1161,7 @@ class About(object):
         if on_windows:
             about_file_path = add_unc(about_file_path)
         with codecs.open(about_file_path, mode='wb', encoding='utf-8') as dumped:
-            dumped.write(self.dumps(with_absent, with_empty))
+            dumped.write(self.dumps(use_mapping, mapping_file, with_absent, with_empty))
 
     def dump_lic(self, location, license_dict):
         """