+++ docs/CHANGELOG.rst
@@ -3,6 +3,10 @@
     Release x.x.x
 
     * Update the list of common license keys
+    * New UrlListField introduced for list of urls
+    * The UrlField is now only taking single URL value
+    * The owner is now a StringField instead of ListField
+    * Format the ordering of the genreated ABOUT file (See https://github.com/nexB/aboutcode-toolkit/issues/349#issuecomment-438871444)
 
 2018-10-23